Move Simple*.ts classes
