+++ src/index.ts
@@ -1,2 +1,4 @@
 export * from './model'
 export * from './ChainOptions'
+export * from './SimpleChain'
+export * from './SimpleProvider'
