+++ test/benchmarks/04-erc20-calls.test.ts
@@ -2,8 +2,7 @@ import { expect } from 'chai'
 import { Wallet } from 'ethers'
 import { Contract } from 'ethers/contract'
 import { parseEther } from 'ethers/utils'
-import { SimpleProvider } from '../../src/chain/SimpleProvider'
-import { createSimpleChain, SimpleChain } from '../../src/chain/SimpleChain'
+import { createSimpleChain, SimpleChain, SimpleProvider } from '../../src'
 import { deployERC20 } from './utils/deploy'
 import { getERC20BalanceOfCall } from './utils/calls'
 import { getERC20TransferTransaction } from './utils/transactions'
